[Priyanshu] : Added Maximum Number Of Handshakes Possible Program

+++ gcr-codebase/methods/Level -1/MaximumHandshakes.java
@@ -1,8 +1,8 @@
 import java.util.Scanner;
 public class MaximumHandshakes{
       // Creating method for calculating the maximum number of handshakes posiible
-       public static double calculatingMaximumHandshakes(double number){
-	          double numberOfHandshakes = (number *(number - 1))/2;
+       public static int calculatingMaximumHandshakes(int number){
+	          int numberOfHandshakes = (number *(number - 1))/2;
                 return numberOfHandshakes;
         }
    
@@ -10,10 +10,10 @@ public class MaximumHandshakes{
               Scanner input = new Scanner(System.in);
               // taking number of students from user
               System.out.print("Enter the Number of Students : ");
-              double numberOfStudents = input.nextDouble();
+              int numberOfStudents = input.nextInt();
               
               // calling method for calculating the maximum number of handshakes posiible 
-              double numberOfHandshakes = MaximumHandshakes.calculatingMaximumHandshakes(numberOfStudents);
+              int numberOfHandshakes = MaximumHandshakes.calculatingMaximumHandshakes(numberOfStudents);
               // printing the results
               System.out.print("The Maximum Number of Possible Handshakes is "+ numberOfHandshakes);
         }
